refactor: BallStorage 제거 및 캐싱 로직 이동
