--- src/main/java/lotto/domain/ball/BallStorage.java
@@ -1,22 +1,0 @@
-package lotto.domain.ball;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class BallStorage {
-
-    private static final Map<Integer, Ball> balls = new HashMap<>();
-
-    public static Ball getBall(final int ballNumber) {
-        saveBallIfNotExist(ballNumber);
-        return balls.get(ballNumber);
-    }
-
-    private static void saveBallIfNotExist(final int ballNumber) {
-        if (balls.containsKey(ballNumber)) {
-            return;
-        }
-        balls.put(ballNumber, new Ball(ballNumber));
-    }
-
-}
