+++ src/main/java/lotto/domain/winning/WinningTicket.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.Optional;
 
 import lotto.domain.ball.Ball;
-import lotto.domain.ball.BallStorage;
 import lotto.domain.rank.Rank;
 import lotto.domain.ticket.Ticket;
 
@@ -16,7 +15,7 @@ public class WinningTicket {
     public WinningTicket(final List<Integer> numbers, final int bonusNumber) {
         WinningTicketValidator.validateWinningTicket(numbers, bonusNumber);
         this.ticket = new Ticket(numbers);
-        this.bonusBall = BallStorage.getBall(bonusNumber);
+        this.bonusBall = Ball.from(bonusNumber);
     }
 
     public Optional<Rank> calculateRank(final Ticket anotherTicket) {
